Remove some unused styl and js files and remove webpack for custom build script.
