--- fix-celery.sh
@@ -1,9 +1,0 @@
-#!/bin/sh
-
-TARGET=venv/lib/python3.7/site-packages/celery/backends
-cd $TARGET
-if [ -e async.py ]; then
-    mv async.py asynchronous.py
-    sed -i 's/async/asynchronous/g' redis.py
-    sed -i 's/async/asynchronous/g' rpc.py
-fi
