--- itdagene/assets/frontend/styl/banner.styl
@@ -1,66 +1,0 @@
-.banner-image
-  width 100%
-  height 500px;
-  padding-top 145px
-
-  .hsp-header
-    color #FFF
-    font-family hsp-font
-    font-size 28px
-    margin-top 5px
-
-  .hsp-logo
-    padding-top 0px
-    margin-top 5px
-
-    img
-      margin: 0 auto
-
-  .location-header
-    color #FFF
-    font-family banner-font
-    font-weight 300
-    font-size 50px
-    padding-top 58px
-    margin-left -28px
-
-  .location-header2
-    color #FFF
-    font-family banner-font2
-    font-size 18px
-    letter-spacing 10px
-    margin-top 0px
-
-  .location-header-small
-    color #FFF
-    font-family banner-font
-    padding-top 20px
-    font-size: 20px
-
-  @media (max-device-width: 640px)
-    .hsp-header
-      font-size 20px
-
-
-.banner-image-small
-  background linear-gradient(rgba(35, 30, 90, 0.4), rgba(50, 35, 90, 0.30)),url("/static/img/banner.jpeg") #1a0a5f no-repeat center center
-  width 100%
-  height 160px
-  background-size cover !important
-  padding-top 90px
-
-  .hsp-header
-    color #FFF
-    font-family hsp-font
-    font-size 28px
-    margin-top 5px
-
-  .hsp-logo
-    img
-      max-height: 65px
-
-  .location-header
-    color #FFF
-    font-family banner-font
-    font-size 40px
-    padding-top 20px
