--- itdagene/assets/frontend/styl/board.styl
@@ -1,20 +1,0 @@
-div.board-holder
-  background-color #FFF
-
-  .member
-    padding-bottom 30px
-
-  h4.member-name
-    font-family header-font
-    margin 10px 0 0 0
-    padding 0
-
-  .member-position
-    font-style italic
-    padding 0
-    margin 5px 0 5px 0
-
-  .member-email
-    padding 0
-    margin 0
-
