--- itdagene/assets/frontend/styl/companies-carousel.styl
@@ -1,41 +1,0 @@
-.companies-holder
-
-  .btn-days
-    border-radius 100px
-    box-shadow none
-    border-color logo-color
-    border-width 2px
-    text-transform uppercase
-
-    &:hover, &:focus, &:active
-      color logo-color
-
-    &:hover
-      color #999
-
-    &.active
-      background-color logo-color
-      color #FFF
-
-  .companies-tab
-    margin-top 20px
-    margin-bottom 20px
-
-  .company-image
-    height 150px
-    display flex
-    align-items center
-
-    img
-      width 100%
-
-
-.collaborators-holder
-  .company-image
-    height 150px
-    display flex
-    align-items center
-
-    img
-      width 100%
-
