--- itdagene/assets/frontend/styl/company.styl
@@ -1,9 +1,0 @@
-@media (min-width 768px)
-  .sidebar
-    border-left dotted #999 2px
-
-h3.header
-    font-family inline-header-font
-    font-size 20px
-.company-contact
-  word-wrap break-word
