--- itdagene/assets/frontend/styl/events.styl
@@ -1,33 +1,0 @@
-#event-page
-  .event-container
-    background linear-gradient(bottom, #EFF5F8, #EFF5F8)
-    border-radius 10px
-
-    .date-title
-      font-family header-font
-      padding-top 15px
-      padding-bottom 0px
-      font-size 24px
-
-    .event-list
-
-      padding-bottom 15px
-
-      li
-        margin-left 15px
-        margin-right 15px
-
-        .item-header
-          display block
-
-          p
-            display inline
-            font-size 18px
-
-        p.event-description
-          color #4287c2
-          font-size 16px
-
-          small
-            color #000
-
