--- itdagene/assets/frontend/styl/footer.styl
@@ -1,51 +1,0 @@
-footer
-  font-family header-font
-
-  &.bottom-menu
-    background-color #262626
-    padding-bottom 20px
-
-    h3.footer-header
-      color #FFF
-      font-size 18px
-
-    address
-      color #98A0A8
-
-    p
-      color #98A0A8
-
-    a
-      color #686c77
-      &:hover
-        color #98A0A8
-        text-decoration none
-
-    i.tweet-date
-      color #7e848b
-
-    .instagram-image
-      padding 2px !important
-
-  &.copyright
-    height 90px
-    background-color #3a3a3a
-
-    p
-      line-height 90px
-      color #98A0A8
-
-      &.copyright-text
-        margin 0
-
-    ul.copyright-menu
-      line-height 90px
-      margin-bottom 0
-
-      li
-        a
-          color #686c77
-
-          &:hover
-            color #98A0A8
-            text-decoration none
