--- itdagene/assets/frontend/styl/itdagene.styl
@@ -1,57 +1,0 @@
-@import 'nib'
-@import "/static/lib/animate.css/animate.min.css"
-@import "/static/lib/font-awesome/css/font-awesome.min.css"
-@import "/static/lib/ionicons/dist/css/ionicons.min.css"
-
-@import url("https://fonts.googleapis.com/css?family=Open+Sans")
-@import url("https://fonts.googleapis.com/css?family=Slabo+27px")
-@import url("https://fonts.googleapis.com/css?family=Lato")
-@import url("https://fonts.googleapis.com/css?family=Marvel")
-@import url('https://fonts.googleapis.com/css?family=Antic');
-
-joblisting-font = 'Antic', sans-serif
-header-font = 'Open Sans', sans-serif
-footer-font = header-font
-hsp-font = 'Open Sans', sans-serif
-banner-font = 'HelveticaNeue-Thin', 'nimbus-sans', sans-serif
-banner-font2 = 'HelveticaNeue-Light', 'nimbus-sans', sans-serif
-inline-header-font = 'Marvel', sans-serif
-logo-color = #4287C2
-
-@import "navbar.styl";
-@import "banner.styl";
-@import "timeline.styl";
-@import "board.styl";
-@import "news.styl";
-@import "joblistings.styl";
-@import "companies-carousel.styl";
-@import "company.styl";
-@import "events.styl";
-@import "footer.styl";
-
-body, html
-  background-color #3a3a3a
-
-a
-  transition 0.3s
-
-.section
-  padding-bottom 20px
-  background-color #FFFFFF
-
-  h2.header
-    font-size 30px
-    word-wrap break-word
-    margin 0
-    padding 0
-    padding-top 30px
-    text-transform uppercase
-
-    &::after
-      width 100px
-      border-bottom 3px solid logo-color
-      margin 4px auto 20px
-      content ""
-      height 0
-      display block
-      position relative
