--- itdagene/assets/frontend/styl/joblistings.styl
@@ -1,53 +1,0 @@
-.news-articles
-
-  .job
-    margin-bottom 20px
-    display block
-
-    .job-header
-      font-family header-font
-      margin 0
-      padding 0
-      color #000000
-      transition 0.3s
-
-    .job-info
-      padding 0
-      margin 0
-
-    &:hover, &:focus, &:active
-      text-decoration none
-
-    &:hover
-      .job-header
-        color logo-color
-
-.joblisting-holder
-
-  h2.header
-    font-family joblisting-font
-    &::after
-      width 100% !important
-
-  h3.header
-    font-family inline-header-font
-    font-size 20px
-
-  h1, h2, h3, h4, h5
-    font-family header-font
-
-  h1
-    font-size 30px
-
-  h2
-    font-size 20px
-
-  h3
-    font-size 16px
-
-  h5, h6
-    font-size 15px
-
-  @media (min-width 768px)
-    .sidebar
-      border-left dotted #999 2px
