--- itdagene/assets/frontend/styl/news.styl
@@ -1,20 +1,0 @@
-.news-articles
-  background linear-gradient(bottom, #EFF5F8, #EFF5F8)
-
-  .news-element
-    color #000
-    display: block
-    margin-bottom 20px
-
-
-    .news-header
-      font-family header-font
-      margin 0
-      padding 0
-
-
-    .news-content
-      font-family footer-font
-
-    &:hover, &:focus, &:active
-      text-decoration none
