--- itdagene/assets/frontend/styl/timeline.styl
@@ -1,37 +1,0 @@
-.hsp-holder
-  background linear-gradient(bottom, #EFF5F8, #EFF5F8)
-
-.timeline-holder
-  .circle-feature
-    height 125px
-    width 125px
-    line-height 135px
-    border-radius 125px
-    border 1px solid #464646
-    text-align center
-    margin auto
-    margin-top 15px
-    margin-bottom 10px
-
-    transition 0.2s
-
-    i.fa
-      font-size 35px
-      color #464646
-
-  h3.feature-header
-    font-family header-font
-    color #464646
-
-
-  p.feature-description
-    font-family header-font
-    color #999
-
-
-  .feature
-
-    &:hover
-      .circle-feature
-        box-shadow 0 0 15px logo-color
-
